Move all parameter validation to ParamsValidator class
The goal here is to separate parameter validation from things like applying defaults and to isolate it so that it's separately testable.

A few testcases need to be moved/rewritten as a result of this commit.  These are now commented out:

test_image_mgmt_upgrade_task_00003
test_image_mgmt_upgrade_task_00004
test_image_mgmt_upgrade_task_00005