ImageUpgradeTask: leverage ParamsMergeDefault class
1. A few boilerplate lines were updated.

2. ImageUpgradeTask
- Previously, we were creating a defaults dictionary which we used to update the merged config.  This was not optimal since we were, in essence, duplicating information about default values for parameters that is already in a properly-written params_spec.

Wrote ParamsMergeDefaults class which takes a params_spec, and a playbook config and updates the playbook config to add any missing parameters that have a default value.

This required re-writing some unit tests.  And, as a result of using this, a errors were found in the params_spec for image_upgrade (the following were not specified: reboot, upgrade.epld, upgrade.nxos).