ImageInstallOptions: Avoid NDFC error
NDFC returns an error when all of issu, epld, and package_install are False.

Avoid this by returning a mocked NDFC response indicating that nothing needs to be done.  This makes the change transparant to anything outside this class.

While we're modifying this class, fix a couple pylint errors and inappropriate assignments of self.method_name.