ImageUpgradeTask: Fix idempotence for merged state
Fix an intermittent idempotence issue found by the integration tests.

have.status was sometimes reported as "Not-In-Sync" which caused upgrade.nxos to be set to True, which cause unnecessary upgrade.  The fix was to remove have.status from consideration.  We now consider only have.reason, have.policy, and have.upgrade.