Add class RestSend(), more...
Replacing ImageUpgradeCommon.dcnm_send_with_retry() with RestSend() class.  The problem with the former has to do with pytest patching.  For some tests, we need to patch dcnm_send() to mock different responses for each of the subclasses of ImageUpgradeCommon.  Since the patch path was identical, there was no way to do this.

The solution is to move this to a separate class (RestSend) which is then imported into subclasses of ImageUpgradeCommon.  This provides the different patch paths that we need to provide appropriate mock responses to each subclass.

Other changes:

1. The above required extensive changes for unit tests of ImageUpgrade (the first subclass we've modified to use RestSend).  Other subclasses will be similarly modified later.

2. Run test_image_upgrade_image_policy_action.py through black/isort

3. InstallOptions:  Move validation of refresh parameters out of refresh() and into a separate method.  Modify unit tests appropriately.

4. InstallOptions: Fix use of self.result, self.result_current, self.response, self.response_current

5. InstallOptions: Remove timeout and unit_test properties (these are moved to RestSend).  Remove response and result properties (these are inherited from ImageUpgradeCommon).  Fix raw_response property to alias response_current.

6. ImageUprade: Use RestSend.

7. ImageUpgradeCommon: Added a lot of debugging to dcnm_send_with_retry()