Make results more like existing module results
Previously, results contained an "OK" and "FAILED" key for each of results, response, diff.

Results for existing modules do not have this.  So, we've changed the results for dcnm_fabric to more closely match the results of existing modules.

We've still retained the unique sequence_number for each of result, response, and diff entries so that they can be correlated.

Also:

Move template_parse_each_fabric.py from module_utils/fabric to module_utils/fabric/vxlan