Fix validate-modules issues
TODO: This is just placeholder documentation.  Need to modify and run a script to generate the actual documetation from the template.  But, this raises an issue... WHICH template?  If dcnm_fabric is supporting multiple fabric types, and they have different templates, which template do we use for the module documentation?