VerifyPlaybookParams(): Remove call to reject_boolean_string() for now.
VerifyPlaybookParams().reject_boolean_string() is rejecting boolean strings in the payload (where they are valid, e.g. FabricReplacedCommon() modifies the payload based on the template, and this modification is inserting boolean strings into the payload).  This method was originally intended to reject boolean strings in the user's playbook only.   Commenting this out for now and may move it later to a method that only acts on the user's playbook.