RuleSet().refresh(): was not resetting self.ruleset prior to updating it.
Caught in integration tests.  The playbook below would fail, as it should.  BUT for the wrong reason.  The reason given was:

"ParamInfo.parameter: Parameter UNDERLAY_IS_V6 not found in fabric template. This means that either the parameter UNDERLAY_IS_V6, or a parameter that is dependent on UNDERLAY_IS_V6, is not appropriate for the fabric type."

The actual reason is that SUBNET_TARGET_MASK is not a valid parameter for LAN_CLASSIC fabric.

The reason that the message is wrong is because RuleSet() was not resetting self.ruleset when refresh() was called.  For the scenario below, this resulted in the following:

1. RuleSet() was updated for VXLAN_EVPN fabric template and the VXLAN_EVPN params were validated.

2. RuleSet() was updated for LAN_CLASSIC fabric template (but not initialized/result, so it still contained the keys from VXLAN_EVPN fabric, which include SUBNET_TARGET_MASK).

As a result of 2, SUBNET_TARGET_MASK parameter in the LAN_CLASSIC fabric playbook config was matched against the remnents of the VXLAN_EVPN template, which resulted in the above misleading error.

The fix is to reset the ruleset in RuleSet().refresh() prior to updating it with the new template.

We now get the correct error for the playbook below, which is:

"FabricUpdateBulk._fabric_needs_update_for_merged_state: Invalid key: SUBNET_TARGET_MASK found in payload for fabric LAN_CLASSIC_Fabric"

    tasks:
        - cisco.dcnm.dcnm_fabric:
            state: merged
            config:
            -   FABRIC_NAME: VXLAN_EVPN_Fabric
                FABRIC_TYPE: VXLAN_EVPN
                BGP_AS: 65535
                DEPLOY: false
            -   FABRIC_NAME: LAN_CLASSIC_Fabric
                FABRIC_TYPE: LAN_CLASSIC
                IS_READ_ONLY: false
                SUBNET_TARGET_MASK: 31
                BOOTSTRAP_ENABLE: false
                DEPLOY: false

Also, cleaned up pylint line-too-long in debug message.