FabricDetails(): Backout last commit (only for FabricDetails)
FabricDetails() has dependent code that does not like the change made in the last commit.

I'll copy FabricDetails() to module_utils/common and make the changes in the copied version.

The existing code within dcnm_fabric can remain as-is and we can modify it to use the new FabricDetails() from module_utils/common later.