sender_file.py Sender(): New class to simulate controller responses
DISCUSSION:

Sender(), in sender_file.py, implements the sender interface.  It is injected into RestSend() v2.  RestSend() has no knowledge of how controller responses are retrieved, as long as the sender interface is conformed to.

Hence, the rest of the code base (that uses RestSend) can be unit tested without having to mock anything.  The actual code is tested, rather than mocks.

1. dcnm_sender.py - rename to sender_dcnm.py for consistency with sender_file.py.  Future Sender() classes, e.g. that leverage Requests, would be named e.g. sender_requests.py, etc.

2. test_maintenance_mode_info.py: Converted all test cases to use the above.

3. MockSwitchDetails(): Modified to mock ONLY the exceptions raised by SwitchDetails.  It no longer duplicates the functionality of Sender().

4. dcnm_maintenance_mode.py: Changed import to use sender_dcnm.

5. module_utils/common/sender_file.py: New file

6. module_utils/common/sender_dcnm.py: Update docstring.

7. module_utils/common/rest_send_v2.py: Update docstring.

8. module_utils/common/maintenance_mode_info.py: Update docstring.