Remove duplicate ResponseGenerator class
ResponseGenerator() was located in both the following locations:

- tests/unit/modules/dcnm/dcnm_fabric/utils.py
- tests/unit/module_utils/common/common_utils.py

We changed RsponseGenerator() to include an "implements" property, which broke all the unit tests that were using the copy that didn't include this property.

Modified all the unit test file imports to point to the copy in common_utils.py and removed the other copy in utils.py.