RestSend() v2: 88% unit test coverage.
1. Added the following test cases

- test_rest_send_v2_00700

Verify ``response.setter`` raises ``TypeError`` when set to inappropriate types, and does not raise when set to dict.

- test_rest_send_v2_00800

Verify ``result_current.setter`` raises ``TypeError`` when set to inappropriate types, and does not raise when set to dict.

- test_rest_send_v2_00900

Verify ``result.setter`` raises ``TypeError`` when set to inappropriate types, and does not raise when set to dict.

2. RestSend() v2: Tweak error messages.