Fix for issue #305
1. FabricReplacedCommon().update_replaced_payload()

If the playbook value for a parameter is None, then the only thing we need to ensure is that the controller value is set to the default value.  If the default value is null, we change it to the empty string "" since NDFC throws a 500 error for null parameter values.

2. Update unit tests to reflect the above change.  The following test was modified.  Specifically two cases where playbook is None (PARAM_8 and PARAM_15).

- test_fabric_replaced_bulk_00040