ImagePolicyAttach(): hardening.
During integration testing, which we updated to do back-to-back upgrades across two different images, we hit a situation where ImagePolicyAttach() started receiving 500 responses from the controller because it was trying to attach an image policy to a switch while one of stage, validate, or upgrade was still in progress.

Added a call to _wait_for_current_actions_to_complete() prior to calling attach_policy().  Hopefully this detects whatever was in progress and avoids the 500 response.