UT: Fix PurePosixPath version differences
Depending on the version of PurePosixPath, the error message differs (see below).  Fix by using a general .* regex for the Error detail: portion of the message.

argument should be a str or an os.PathLike object where __fspath__ returns a str, not 'int'

Versus:

expected str, bytes or os.PathLike object, not int'