dcnm_fabrics_vrfs: remove get_have_nv()
get_have_nv(): This method was used only for testing the corresponding class EppFabricsVrfsByKeyValue().  Since we are not using this class in this module, and basic testing for the class is finished, we're removing this code.

epp_vrfs.py: Updating usage docstrings to reflect the path changes we made in the last commit.