WIP: epp_sites.py: Add more properties
The sites endpoint returns a complex structure with several dicts.  Created a generic method to extract values for these dicts -- given a property that returns the dict, and a keyname within the dict -- and leveraged it in properties to return individual dict item values.

Still a work in progress since there are several more dicts to process.