dcnm_fabric: cleanup rule_value and user_value processing
Prior to passing eval_string to eval() we need to ensure that rule_value and user_value are not null (None, "", or "null), else, we'll get a syntax error in eval().

Change the handling here such that:

if rule_value is in [None, "", "null"], return True (since, in this case, there is no rule for the parameter).

if user_value is in [None, "", "null"] set the user_value such that the eval() will pass (given current value of operator).

Also, modifications based on pylint recommendations.