More refactoring
1. update_attach_params()

Simplified by:

1. Initialize nbr_dict values to ""
2. Populate the nbr_dict values from the current item from attach["vrf_lite"]
3. Test if any values in nbr_dict are != "" after step 2
4. If no values have been updated, continue
5. De-indent the remaining code
6. (also renamed vlanId to vlan_id)

This change also required that we add "IF_NAME" to self.vrf_lite_properties.  I verified that this change will not impact the other use for this structure in diff_for_attach_deploy().

2. diff_for_create()

After the refactor of this method in the last commit, it became obvious that code in the if and else clauses were heavily duplicated.

Refactored to remove the if/else clause entirely since the only difference between these was whether we skip key "vrfVlanId" when vlan_id_want == 0.

This reduces to a simple if statement to populate skip_keys if vlan_id_want == 0.