Remove template_name validation
EpTemplate() was using FabricTypes() to validate the template_name property against a list of valid and supported fabric templates.  However, the intent of EpTemplate() is to retrieve other template types as well (network, vrf, etc), so this validation isn't ideal in this larger context.

This commit removes this validation and the associated import and instantiation of FabricTypes.

There may be associated unit-tests that need to be updated.  Will investigate and handle in a second commit if so.