UT: Remove test_ep_templates_00050
1. Remove test_ep_templates_00050()

This test validated that a ValueError is raised if template_name is invalid.  The test for validity entailed comparing against a list of fabric template names in FabricTypes().  However, EpTemplate() is now responsible for retrieving templates that are not in this list (e.g. interface templates, etc).  Hence, this test is no longer valid.