EpTemplate() - remove template_name validation (#355)
* Remove template_name validation

EpTemplate() was using FabricTypes() to validate the template_name property against a list of valid and supported fabric templates.  However, the intent of EpTemplate() is to retrieve other template types as well (network, vrf, etc), so this validation isn't ideal in this larger context.

This commit removes this validation and the associated import and instantiation of FabricTypes.

There may be associated unit-tests that need to be updated.  Will investigate and handle in a second commit if so.

* UT: Remove test_ep_templates_00050

1. Remove test_ep_templates_00050()

This test validated that a ValueError is raised if template_name is invalid.  The test for validity entailed comparing against a list of fabric template names in FabricTypes().  However, EpTemplate() is now responsible for retrieving templates that are not in this list (e.g. interface templates, etc).  Hence, this test is no longer valid.