Initial commit
1. module_utils/network/dcnm.py

Add two new methods which are duplicate functionality to get_ip_sn_fabric_dict()

get_ip_fabric_dict()
get_sn_fabric_dict()

get_ip_sn_fabric_dict() was not touched.

The reason for the new defs is because patch (in test_dcnm_vrf.py setUp() did not like patching to a def that returns more than one value.  But another reason is that it's cleaner for these utility defs to be as simple as possible.

2. test_dcnm_vrf.py

- Add the mock for sn_fab

- Modify load_fixtures() for all test cases that required having sn_fab dict mocked.

- Define two new vars for fabric details (for a future commit)
    -  fabric_details_mfd
    - fabric_details_vxlan

3. dcnm_vrf.json

Add the following objects (fabric_details_* for a future commit)

- mock_sn_fab
- fabric_details_mfd
- fabric_details_vxlan