dcnm_log: Appease linters
1. Update module documentation

Apparently argument_spec.choices is case sensitive.  So, for example, "debug" will not work as a value for severity unless it is explicitely added to the choices list.  It's simpler just not to support case-insensitivity.

Modified the documentation to be explicit that UPPERCASE e.g. DEBUG rather than debug, is mandatory for the dcnm_log severity parameter.

2. Run through linters black, isort, pylint.

3. Modify integration tests to remove positive test for lowercase debug severity parameter.

4. Update tests/sanity/ignore-*.txt with missing-gplv3-license exception for dcnm_log.py