dcnm_log: Module to log from a playbook into the file pointed to by to NDFC_LOGGING_CONFIG (#367)
* Initial commit

1. plugins/modules/dcnm_log.py

An Ansible module for the DCNM Collection to log to the file pointed to by NDFC_LOGGING_CONFIG.

2. tests/integration/targets/dcnm_log/*

Integration tests for dcnm_log.py

* dcnm_log: Appease linters

1. Update module documentation

Apparently argument_spec.choices is case sensitive.  So, for example, "debug" will not work as a value for severity unless it is explicitely added to the choices list.  It's simpler just not to support case-insensitivity.

Modified the documentation to be explicit that UPPERCASE e.g. DEBUG rather than debug, is mandatory for the dcnm_log severity parameter.

2. Run through linters black, isort, pylint.

3. Modify integration tests to remove positive test for lowercase debug severity parameter.

4. Update tests/sanity/ignore-*.txt with missing-gplv3-license exception for dcnm_log.py

* dcnm_log: "message" is not a valid playbook parameter. Change to log_message

1. plugins/modules/dcnm_log.py

Ansible sanity fails because the parameter "message" is used internally by Ansible.  Hence, we need a different parameter name to specific the message to log.  For now, am changing this to log_message.

Updated code and documentation to use log_message instead of message.

2. Updated integration tests to use log_message instead of message.

* Address sanity errors. Use "msg" instead of "log_message"

1. Fix Ansible sanity errors.

2. Decided to use msg rather than log_message to specify the message to log.

3. Rename element_spec to argument_spec which is more in line with Ansible terminology.

* dcnm_log: Fix sanity errors part 2, add integration test

1. Fix typo in argument_spec that was overwriting argument_spec.msg.required value.

2. Add negative integration test for missing msg specification.

* dcnm_log: run collection_prep to generate docs for dcnm_log.py

1. Update and run collection_prep

This updated the following:

1. README.md
2. docs/cisco.dcnm.dcnm_maintenance_mode_module.rst (fixed the version added field)

And added the following:

1. docs/cisco.dcnm.dcnm_log_module.rst

* dcnm_log: Avoid passing Ansible module to DcnmLog class

Limit the use of AnsibleModule() to main() by passing only module.params to DcnmLog().

* Add unit tests, add integration test, update copyright dates

1. Add usit tests for basic coverage

2. Add negative integration test for missing msg parameter

2. Update copyright dates in all files

* Appease yamllint: Remove extra blank line in test_logging.yaml

* Appease pylint. Rename unit tests.

1. plugins/modules/dcnm_log.py

Fix pylint used-before-assignment error.

2. tests/unit/modules/dcnm_log/test_dcnm_log.py

Rename all test cases from test_log_* to test_dcnm_log_*