Issue #352 : more preparatory changes
1. plugins/modules/dcnm.vrf

- send_to_controller()

Add a log_response parameter which controls whether the controller response will be added to the ansible response.

- get_have() - simplify conditional

attach_state = False if attach["lanAttachState"] == "NA" else True

Changed to:

attach_state = not attach["isLanAttached"] == "NA"

- release_resources_by_id()

Fix pylint error (id_list was being redefined in the for loop, so renamed to item).

- failure(): disable pylint check for protected-access

2. tests/integration/targets/dcnm_vrf/tests/dcnm/merged.yaml

There are no functional changes here.  We've only done the following:

- Remove single-quotes from all asserts
- Remove jinja2 templating around variables in asserts (these generate warnings and are not needed).
- Removing the single-quotes broke asserts of the following form:

- "test phrase" in var

To fix this, added set fact for "test phrase" like so:

- name: set fact
  set_fact:
    TEST_PHRASE: "test phrase"

- assert:
  that:
    - TEST_PHRASE in var

3. plugins/__init__.py

- Added to appease pylint