dcnm_vrf: Restructure DcnmVrf to isolate DCNM v11 code
This is a large commit.  Sorry!  I don't see a way to do this without updating everything at once.

1. Deleting unit tests (will add them in the next commit.

Doing this so that git doesn't think we've modified a file. This will make the diffs shorter.

2. Adding separate files in module_utils/vrf for the main module code

- plugins/module_utils/dcnm_vrf_v11.py
- plugins/module_utils/dcnm_vrf_v12.py

3. The main dcnm_vrf.py file now consists of main() and a stub DcnmVrf() class that serves as a launcher for one of the above two files, depending on the controller version.

4. VrfControllerToPlaybookModel also broken up into v11 and v12 versions.

- module_utils/vrf/vrf_controller_to_playbook_v11.py
- module_utils/vrf/vrf_controller_to_playbook_v12.py

We will delete the existing vrf_controller_to_playbook.py file in the next commit.  This is to make the diffs shorter (same reason as item 1 above).

5. VrfPlaybookModel, like item 4, broken up into v11 and v12 versions.

- plugins/module_utils/vrf_playbook_model_v11.py
- plugins/module_utils/vrf_playbook_model_v12.py

We will delete the existing vrf_playbook_model.py file in the next commit.  This is to make the diffs shorter (same reason as items 1 and 4 above).