dcnm_vrf: get_have() fix skipping of attachments
In get_have(), we were returning if any attachment in the attach_list didn’t contain lite_objects.  This likely impacted (at least)  idempotence cases since it would indication that there were no attachments when, in fact, there were.

I’ve left the original code (the return statement) commented out for now.

Unit tests are passing with this change, but this section of code is never hit in the unit tests (which indicates a hole in our testing).