handle_response(): update docstring
Update the docstring with description of parameters and return values.

NOTE: In reviewing this method, I see that the “query” action handler returns chenged = True, which should never be the case.

We need to review this case.

Indeed, this method should be rewritten since the return values are overloaded by the callers of this method.  For example:

The following interpret fail as missing_fabric and changed as not_ok

- get_vrf_objects() interprets
- get_next_vrf_id
- get_diff_query

The following interpret fail as fail, and changed as changed:

- diff_merge_create
- send_to_controller
- release_orphan_resource

However, making any changes here causes unit tests to fail, so I think we need to pair program a solution.