dcnm_vrf: IT: Model tweaks
Integration tests were failing due to model validation errors.  Tweaked several models to address this.

Detail:

In several cases, NDFC can return either an int or None.  It was initially unclear to me how Pydantic handles validation for this sort of union of types.  I wrote a test script to understand the behavior empirically, and arrived at the following (the below is Pydantic default behavior as of version 2, referred to in the Pydantic documentation as 'smart mode'), here:

https://docs.pydantic.dev/latest/concepts/unions/

Assume the following model:

class NoneIntModel(BaseModel):
    value: Union[None, int] = Field(default=None, ge=2, le=4094)

If the input is None, the validation (2 < x >=4094) does not take place.
If the value is an int, the validation does take place.
If the value is some other type, (other than int, None) validation is perform against the type resulting in an (expected) error.

So, for now, this seems to be the way to handle these sorts of controller responses.

Models changed:

- plugins/module_utils/vrf/controller_response_vrfs_attachments_v12.py
- plugins/module_utils/vrf/controller_response_vrfs_switches_v12.py
- plugins/module_utils/vrf/controller_response_vrfs_v12.py

This also means that we can probably simplify the VrfPayloadV12 model in:

- plugins/module_utils/vrf/vrf_controller_payload_v12.py

Will look into this later.