1. plugins/module_utils/vrf/dcnm_vrf_v12.py
- update_vrf_template_config_from_model() new method

Temporary new method which is functionally equivilent to update_vrf_template_config() but takes model ControllerResponseVrfsV12() and returns a stringafied vrfTemplateConfig.

- update_vrf_model_from_dict() - new method.  UNUSED.  Remove later…

- self.get_vrf_objects()

Modify temporarily to return the original dictionary AND a model.  We will later remove the original dictionary from this method’s return.  For now, we need both, since this is called from a couple places and we are updating these callers to accept a model instead of a dict.

- get_diff_query()

Modify the call to get_vrf_objects to accept both a dict and a model

- get_have()

Modify the call to get_vrf_objects to accept both a dict and a model.

Leverage update_vrf_template_config_from_model()