Another experimental commit, since the last commit resulted in integration test failures.
We need a way to convert a VrfObject model into a VRF payload (which contains vrfTemplateConfig as a JSON string rather than a model or dict).

This second attempt is to modify VrfControllerPayloadV12 (which has not been used up until now) to add a field_serializer which converts vrfTemplateConfig to a JSON string when dumping the model.

We leverage this initially in push_diff_create() where we validate VRF objects using VrfObjectV12.  We then dump VrfObjectV12 into VrfPayloadV12 and then dump VrfPayloadV12 to form the actual payload to be sent to the controller.

1. plugins/module_utils/vrf/vrf_controller_payload_v12.py

Add a field_serializer to convert vrfTemplateConfig to a JSON string when dumping the model.

2. plugins/module_utils/vrf/dcnm_vrf_v12.py

- push_diff_create() modified to send a proper payload to the controller.

This method currently pushes each VRF in self.diff_create (list of VRF dicts) to the controller.  Eventually self.diff_create will be a list of VrfPayloadV12 models, but we’re not there yet.  For now, we convert each vrf object in self.diff_create to a VrfObjectV12 model, pass this model to update_vrf_template_config_from_vrf_template_model() where two fields in vrfTemplateConfig are updated, then convert the updated model to a VrfPayloadV12 model, which is used as the final payload.