Experimental: convert to JSON as late as possible
- update_vrf_template_config_from_vrf_model()

Originally, this method accepted a VrfObjectV12 model and returned a dict containing the modified vrfTemplateConfig field.

Our overall goal is to convert inputs to models as early as possible, work on those models, and then output the models as JSON as late as possible.  In this spirit, we’ve modified this method to accept a model and return a model.

No changes were required to the methods that call this method.

We’ve commented out some code which we’ll remove in the next commit if integration tests pass.