get_diff_query: refactor
1. get_diff_query

The main if-else clause of the original method handles two cases for VRFs in self.fabric

1. if - Create a query-state diff for VRFs present in want
2. else - Create a query-state diff for all VRFs

Refactor this into two methods, called from get_diff_query, that perform the actions described above.

We are retaining the original get_diff_query code (commented-out) so that we can easily revert if integration tests fail.

We will clean this up in the next commit.