update_lan_attach_list: return pydantic model
NOTE: We are retaining the original methods (renamed to *_orig) until we’ve verified that integration tests pass.

This commit infuses Pyydatic deeper into the dcnm_vrf’s methods.

The methods below are all within class NdfcVrfV12.

1. update_lan_attach_list

- call get_vrf_lite_objects_model instead of get_vrf_list_objects

- operate on the list[ExtensionPrototypeValue] (rather than original list[dict])

- pass lite (now a list of ExtensionPrototypeValue) to update_vrf_attach_vrf_lite_extensions

2. update_vrf_attach_vrf_lite_extensions

- Modify to work with list of ExtensionPrototypeValue

- call get_extension_values_from_lite_objects with list of  ExtensionPrototypeValue

- get_extension_values_from_lite_objects  now returns a list of VrfLiteConnProtoItem

- Use this list of VrfLiteConnProtoItem to update nbr_dict

3. get_extension_values_from_lite_objects

This method now operates on Pydantic models exclusively.

4. log_list_of_models

It is anticipated that we will be working with lists of models extensively.

This utility method logs a list of pydantic models by calling json.dumps(model.model_dump()) on each model in the list.

It is currently called from all of the above methods.