Fix two unit tests, replace dcnm_get_url
1. plugins/module_utils/network/dcnm/dcnm.py

No changes were made to this file.  However, the function
dcnm_get_url is not a generic function and so should be
moved under module_utils/vrf (IMHO).

We’ve left the original method, since it it still used by:

- dcnm_vrf_v11.py
- test_dcnm_vrf_v11.py

We have rewritten and renamed dcnm_get_url and now use the rewritten version in the following files:

- plugins/module_utils/vrf/dcnm_vrf_v12.py
- tests/unit/modules/dcnm/test_dcnm_vrf_12.py

The rewritten version is renamed to get_endpoint_with_long_query_string and is located here:

- plugins/module_utils/vrf/vrf_utils.py

2. While debugging this new function, it was observed that the following unit test was passing, but it was passing for the wrong reason:

- test_dcnm_vrf_12_merged_redeploy

The code tested by the above unit test calls dcnm_get_url, but there was no side_effect defined for it e.g.:

self.run_dcnm_get_url.side_effect = [self.mock_vrf_attach_object_pending]

Because of this, the controller response was a string containing something along the lines of “<Mock id=(memory_address)>” rather than an actual mocked response.

After adding the above side_effect, the test is now passing for (hopefully) the right reason, and the controller response is as expected.

3. plugins/module_utils/vrf/dcnm_vrf_v12.py

- Remove import for dcnm_get_url

- Add import for get_endpoint_with_long_query_string

- get_have
  - modify to use get_endpoint_with_long_query_string
  - Validate the controller response with a Pydantic model

We don’t currently pass the model to populate_have_deploy and populate_have_attach.  We’ll do that in a separate commit.