push_diff_detach: fabricName -> fabric
1. push_diff_detach

Changes in an earlier commit appear to have caused self.diff_detach to contain fabricName key instead of fabric key.

If fabricName is is found in any of the items in self.diff_detach.lanAttachList, replace it with fabric key.

We need to go back and determine where this is getting changed to fabricName (probably in one of the models).  Once we figure this out, we’ll revert the changes in this commit.