Revert last commit
Revert field_validator for bgp_passwd_encrypt since it causes an Internal Server Error on NDFC.

Ideally, bgpPasswordKeyType would always be in int() or str(), and in ND4.x, it IS always a str().  Unfortunately, that’s not going to help in ND 3.x where we will have to accept Union[int, str].

For now, I’m just reverting the last commit.  Will think about a better solution…